+++ packages/file/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/proteinjs/db/compare/@proteinjs/db-file@1.3.0...@proteinjs/db-file@1.3.1) (2026-07-10)
+
+**Note:** Version bump only for package @proteinjs/db-file
+
+
+
+
+
 # [1.3.0](https://github.com/proteinjs/db/compare/@proteinjs/db-file@1.2.49...@proteinjs/db-file@1.3.0) (2026-06-19)